requirements.txt
# Python
pytest==8.2.2
pillow==10.4.0
pytest-cov==5.0.0

# Debug
ipdb
